+++ src/items/mod.rs
@@ -3,7 +3,6 @@ pub mod menu_item;
 pub use menu_item::MenuItem;
 
 use embedded_graphics::{
-    draw_target::DrawTarget,
     mono_font::MonoTextStyle,
     pixelcolor::BinaryColor,
     prelude::{Point, Size},
@@ -13,6 +12,8 @@ use embedded_graphics::{
 };
 use embedded_layout::prelude::*;
 use embedded_text::{alignment::HorizontalAlignment, style::TextBoxStyleBuilder, TextBox};
+
+use crate::adapters::Canvas;
 
 /// Marker trait necessary to avoid a "conflicting implementations" error.
 pub trait Marker {}
@@ -33,13 +34,49 @@ pub trait MenuListItem<R>: Marker + View {
         true
     }
 
-    fn draw_styled<D>(
+    fn draw_styled(
         &self,
         text_style: &MonoTextStyle<'static, BinaryColor>,
-        display: &mut D,
-    ) -> Result<(), D::Error>
-    where
-        D: DrawTarget<Color = BinaryColor>;
+        display: &mut dyn Canvas<BinaryColor>,
+    ) -> Result<(), ()>;
+}
+
+impl<R> Marker for &mut dyn MenuListItem<R> {}
+
+impl<R> View for &mut dyn MenuListItem<R> {
+    fn translate_impl(&mut self, by: Point) {
+        (**self).translate_impl(by)
+    }
+
+    fn bounds(&self) -> Rectangle {
+        (**self).bounds()
+    }
+}
+
+impl<R> MenuListItem<R> for &mut dyn MenuListItem<R> {
+    fn value_of(&self) -> R {
+        (**self).value_of()
+    }
+
+    fn interact(&mut self) -> R {
+        (**self).interact()
+    }
+
+    fn set_style(&mut self, text_style: &MonoTextStyle<'_, BinaryColor>) {
+        (**self).set_style(text_style)
+    }
+
+    fn selectable(&self) -> bool {
+        (**self).selectable()
+    }
+
+    fn draw_styled(
+        &self,
+        text_style: &MonoTextStyle<'static, BinaryColor>,
+        display: &mut dyn Canvas<BinaryColor>,
+    ) -> Result<(), ()> {
+        (**self).draw_styled(text_style, display)
+    }
 }
 
 /// Helper struct to draw a menu line that has a title and some additional marker.
@@ -72,17 +109,14 @@ impl MenuLine {
         }
     }
 
-    pub fn draw_styled<D>(
+    pub fn draw_styled(
         &self,
         title: &str,
         value_text: &str,
         text_style: &MonoTextStyle<'static, BinaryColor>, // TODO: allow non-mono fonts
-        display: &mut D,
-    ) -> Result<(), D::Error>
-    where
-        D: DrawTarget<Color = BinaryColor>,
-    {
-        let display_area = display.bounding_box();
+        mut display: &mut dyn Canvas<BinaryColor>,
+    ) -> Result<(), ()> {
+        let display_area = display.bounds();
 
         if self.bounds.intersection(&display_area).size.height == 0 {
             return Ok(());
@@ -101,10 +135,10 @@ impl MenuLine {
                 .alignment(HorizontalAlignment::Right)
                 .build(),
         )
-        .draw(display)?;
+        .draw(&mut display)?;
 
         text_bounds.size.width -= self.value_width;
-        TextBox::new(title, text_bounds, *text_style).draw(display)?;
+        TextBox::new(title, text_bounds, *text_style).draw(&mut display)?;
 
         Ok(())
     }
@@ -118,4 +152,10 @@ impl View for MenuLine {
     fn bounds(&self) -> Rectangle {
         self.bounds
     }
+}
+
+#[cfg(test)]
+mod test {
+    #[allow(unused)]
+    fn _is_object_safe(_: &dyn super::MenuListItem<u32>) {}
 }